Merge dev into patch-22

+++ rtgui/coordinateadjuster.cc
@@ -69,14 +69,14 @@ void CoordinateAdjuster::AxisAdjuster::setValue(double newValue)
 {
     float range = rangeUpperBound - rangeLowerBound;
     spinButtonConn.block(true);
-    spinButton->set_value(newValue * range + rangeLowerBound);
+    spinButton->set_value(static_cast<float>(newValue) * range + rangeLowerBound);
     spinButtonConn.block(false);
 }
 
 void CoordinateAdjuster::AxisAdjuster::valueChanged()
 {
     float range = rangeUpperBound - rangeLowerBound;
-    parent->updatePos(idx, (spinButton->get_value() - rangeLowerBound) / range);
+    parent->updatePos(idx, (static_cast<float>(spinButton->get_value()) - rangeLowerBound) / range);
 }
 
 CoordinateAdjuster::CoordinateAdjuster(CoordinateProvider *provider, CurveEditorSubGroup *parent, const std::vector<Axis> &axis)
@@ -171,7 +171,7 @@ void CoordinateAdjuster::startNumericalAdjustment(const std::vector<Boundaries> 
         Gtk::SpinButton *currSpinButton = axisAdjusters.at(i)->spinButton;
         currSpinButton->set_sensitive(true);
         float range = axisAdjusters.at(i)->rangeUpperBound - axisAdjusters.at(i)->rangeLowerBound;
-        currSpinButton->set_range(newBoundaries.at(i).minVal * range + axisAdjusters.at(i)->rangeLowerBound, newBoundaries.at(i).maxVal * range + axisAdjusters.at(i)->rangeLowerBound);
+        currSpinButton->set_range(newBoundaries.at(i).minVal * static_cast<double>(range) + static_cast<double>(axisAdjusters.at(i)->rangeLowerBound), newBoundaries.at(i).maxVal * static_cast<double>(range) + static_cast<double>(axisAdjusters.at(i)->rangeLowerBound));
     }
 
     axisAdjusters.at(0)->spinButton->grab_focus();
@@ -200,7 +200,7 @@ void CoordinateAdjuster::switchAdjustedPoint(std::vector<double> &pos, const std
 
         // ...narrow the range to the new interval
         float range = axisAdjusters.at(i)->rangeUpperBound - axisAdjusters.at(i)->rangeLowerBound;
-        currAxis->spinButton->set_range(newBoundaries.at(i).minVal * range + axisAdjusters.at(i)->rangeLowerBound, newBoundaries.at(i).maxVal * range + axisAdjusters.at(i)->rangeLowerBound);
+        currAxis->spinButton->set_range(newBoundaries.at(i).minVal * static_cast<double>(range) + static_cast<double>(axisAdjusters.at(i)->rangeLowerBound), newBoundaries.at(i).maxVal * static_cast<double>(range) + static_cast<double>(axisAdjusters.at(i)->rangeLowerBound));
 
         // enable events
         currAxis->spinButtonConn.block(false);
